Feature/wrapper updates (#2)
* feat(Stagger): Add onComplete callback

Allow users to pass an  `onComplete` function that will be called after the last child transition is
complete

BREAKING CHANGE: the `duration` prop now must be passed to `Stagger` so it can calculate the correct
timing for `onComplete`. Any `duration` in child animations will be overwritten when wrapped in
`Stagger`

* feat(Random): Add onComplete callback

Allow users to pass an onComplete function that will be called after the last transition completes

BREAKING CHANGE: Users must pass `duration` into `Random` now rather than directly into the child
animation component. This is so it can properly calculate when to fire the `onComplete` callback

* refactor(Storybook): Update stories to use `map`

Create one array of `Example` text and map through it in each Wrapper example

* feat(Stagger): Add `in` prop for conditionally starting Stagger

Users can pass an `in` prop to `Stagger` which will mount or unmount the children and cause them to
Stagger in and out

BREAKING CHANGE: Stagger is no longer set to `appear` by default. You must pass `in` as true to
initiate

* feat(Random): Add `in` prop for conditionally starting Random

Users can pass `in` to Random to mount and unmount the child animations similayly to `Stagger`

BREAKING CHANGE: Random is no longer set to `appear`. Must pass `in` as `true` to initiate the
Animation