Feature/animation updates (#3)
* feat(Animations): Allow classNames

Allow a user to pass a className to the inner wrapping div on each animation.

* test(Snapshots): Update snapshots for `className`

Rendered element snapshot no includes `className={undefined}`

* feat(FadeInOut): Add `enterOpacity` and `exitOpacity` props

All a user to set what the enter and exit values for opacity are on `FadeInOut`. Defaults to enter
as 1 and exit as 0

* fix(TweenTransform): Fix collision with prop names

`enter` and `exit` are `Transition` props. Change to `enterTransform` and `exitTransform`

BREAKING CHANGE: `enter` and `exit` will now be passed to the transition as a prop and will not
control the values for transforms

* refactor(FadeTransform): Make FadeTransform a composed component

Rather than a separate API for FadeTransform, use both existing animation components to create a
composed version. Allows a user to pass shared props for transition values an individual prop
objects for each animation

BREAKING CHANGE: The API changed to accept shared props at the parent level and props for individual
inner animation components in `fadeProps` and `transformProps`

* refactor(Animations): Rename animations

Rename animation files and references to be shorter and more concise names

BREAKING CHANGE: Old imports will no longer work as the components are exported with their new names

* test(Snapshots): Update snapshots for renamed components