New Features and Bug Fixes (#6)
* Feature/wrapper updates (#2)

* feat(Stagger): Add onComplete callback

Allow users to pass an  `onComplete` function that will be called after the last child transition is
complete

BREAKING CHANGE: the `duration` prop now must be passed to `Stagger` so it can calculate the correct
timing for `onComplete`. Any `duration` in child animations will be overwritten when wrapped in
`Stagger`

* feat(Random): Add onComplete callback

Allow users to pass an onComplete function that will be called after the last transition completes

BREAKING CHANGE: Users must pass `duration` into `Random` now rather than directly into the child
animation component. This is so it can properly calculate when to fire the `onComplete` callback

* refactor(Storybook): Update stories to use `map`

Create one array of `Example` text and map through it in each Wrapper example

* feat(Stagger): Add `in` prop for conditionally starting Stagger

Users can pass an `in` prop to `Stagger` which will mount or unmount the children and cause them to
Stagger in and out

BREAKING CHANGE: Stagger is no longer set to `appear` by default. You must pass `in` as true to
initiate

* feat(Random): Add `in` prop for conditionally starting Random

Users can pass `in` to Random to mount and unmount the child animations similayly to `Stagger`

BREAKING CHANGE: Random is no longer set to `appear`. Must pass `in` as `true` to initiate the
Animation

* Feature/animation updates (#3)

* feat(Animations): Allow classNames

Allow a user to pass a className to the inner wrapping div on each animation.

* test(Snapshots): Update snapshots for `className`

Rendered element snapshot no includes `className={undefined}`

* feat(FadeInOut): Add `enterOpacity` and `exitOpacity` props

All a user to set what the enter and exit values for opacity are on `FadeInOut`. Defaults to enter
as 1 and exit as 0

* fix(TweenTransform): Fix collision with prop names

`enter` and `exit` are `Transition` props. Change to `enterTransform` and `exitTransform`

BREAKING CHANGE: `enter` and `exit` will now be passed to the transition as a prop and will not
control the values for transforms

* refactor(FadeTransform): Make FadeTransform a composed component

Rather than a separate API for FadeTransform, use both existing animation components to create a
composed version. Allows a user to pass shared props for transition values an individual prop
objects for each animation

BREAKING CHANGE: The API changed to accept shared props at the parent level and props for individual
inner animation components in `fadeProps` and `transformProps`

* refactor(Animations): Rename animations

Rename animation files and references to be shorter and more concise names

BREAKING CHANGE: Old imports will no longer work as the components are exported with their new names

* test(Snapshots): Update snapshots for renamed components

* feat(Storybook): Add more controls

Add `chunk` option in `Stagger` and `minDelay` and `maxDelay` in `Random`

* chore(Eslint): Add eslintrc

* refactor(Storybook): Split stories into separate files. Add `createCommonKnobs` to utilities

* chore(Linting): Code clean up

Satisfy eslint rules

* feat(Wrappers): Implement Reverse (#4)

Allow users to pass `reverse` to reverse the way the delays are applied in `Stagger` and `Random`.
Also includes a bug fix where `onComplete` was not called `onExited` and was being called multiple
times

* docs(Readme): Update readme for new features

* docs(Readme): Fix section links

* docs(Readme): Fix typo

* docs(Readme): Fix typo

* test(Unit Tests): Clean up unneeded imports. Add test globals to eslintrc

* test(Snapshot): Add snapshot tests for passing className

* test(Snapshot): Fix snapshots for updates

* docs(Readme): Fix subtitle

* docs(Readme): More concise subtitle